TRE-106: chore(stats): refactor

+++ core/network/src/main/java/com/doannd3/treetask/core/network/model/response/StatsResponse.kt
@@ -10,7 +10,7 @@ data class DashboardResponse(
     @SerialName("inProgress") val inProgress: Int? = null,
     @SerialName("pending") val pending: Int? = null,
     @SerialName("done") val done: Int? = null,
-    @SerialName("completionRate") val completionRate: Int? = null,
+    @SerialName("completionRate") val completionRate: Double? = null,
     @SerialName("recentTasks") val recentTasks: List<RecentTask> = emptyList(),
 )
